+++ src/town.h
@@ -84,10 +84,8 @@ struct Town : TownPool::PoolItem<&_town_pool> {
 
 	uint16 time_until_rebuild;       ///< time until we rebuild a house
 
-	uint32 grow_counter;             ///< counter to count when to grow, value is smaller than or equal to growth_rate
-
-	// Stepan: In might be quite a big number.
-	uint32_t growth_rate;              ///< town growth rate
+	uint16 grow_counter;             ///< counter to count when to grow, value is smaller than or equal to growth_rate
+	uint16 growth_rate;              ///< town growth rate
 
 	byte fund_buildings_months;      ///< fund buildings program in action?
 	byte road_build_months;          ///< fund road reconstruction in action?
